Another fix for JSDocs for nixysa.
It turns out the JSCompiler is fine with.

/**
 * Some property.
 * @type {number}
 */
MyClass.prototype.someProperty;

But the JSCompiler needs

/**
 * Some property.
 * @type {number}
 */
MyClass.prototype.someProperty = something;

otherwise it doesn't pick up the properties.

The JSCompiler would not be fine with that because the type on
the right side of the equals would have to match the type.


Review URL: http://codereview.chromium.org/149715